requirements.txt
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.5
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
frozenlist==1.5.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
jiter==0.9.0
joblib==1.4.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.3.2
packaging==24.2
platformdirs==4.3.7
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.2.3
pypiwin32==223
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-vlc==3.0.21203
pyttsx3==2.90
pytube==15.0.0
pywin32==306
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.10.4
srt==3.5.3
toml==0.10.2
tqdm==4.66.4
urllib3==2.2.2
virtualenv==20.30.0
websockets==15.0.1
yarl==1.19.0
yt-dlp==2025.3.31
